requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.8
django-crispy-forms==2.1
django-phonenumber-field==7.2.0
django-recaptcha==4.0.0
phonenumbers==8.13.26
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
